Turtle Crossing game
This is an Animal Crossing- like game, which is built using Python turtle module. The user controlling a turtle has to cross a path on which there are cars assing. Once a car hits the turtle- game is over. When user gets to the other side of the path, a next level begins with a higher speed of the cars.